dad-is-great-in-C: less includes

+++ dad-is-great-in-C/lastword.cpp
@@ -1,6 +1,5 @@
 // task: print last word of input
 
-#include <cstdint>
 #include <iostream>
 #include <stddef.h>
 #include <string>
@@ -10,8 +9,8 @@ int main() {
   std::getline(std::cin, line);
   size_t end = line.find_last_not_of(" ");
   std::string trimmed = line.substr(0, end + 1);
-  std::string last(trimmed.begin() + static_cast<int64_t>(trimmed.rfind(" ")) +
-                       1,
+  std::string last(trimmed.begin() +
+                       static_cast<long long>(trimmed.rfind(" ")) + 1,
                    trimmed.end());
   std::cout << last;
   return 0;
